xauusd-predictive-analytics/requirements.txt
# ── Core connector ────────────────────────────────────────────────────────────
MetaTrader5>=5.0.4424

# ── Data processing ───────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0

# ── Technical analysis ────────────────────────────────────────────────────────
pandas-ta>=0.3.14b
TA-Lib>=0.6.0

# ── AI / ML model ─────────────────────────────────────────────────────────────
catboost>=1.2.0
scikit-learn>=1.3.0

# ── Logging ───────────────────────────────────────────────────────────────────
loguru>=0.7.0

# ── Configuration ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
